+++ omni/entrypoints/start_comfyui.sh
@@ -1,0 +1,5 @@
+export http_proxy=<your_http_proxy>
+export https_proxy=<your_https_proxy>
+export no_proxy=localhost,127.0.0.1
+
+python /llm/ComfyUI/main.py --listen 0.0.0.0 --port 8188